+++ tests/helper/helper_oc.go
@@ -274,7 +274,7 @@ func (oc *OcRunner) ServiceInstanceStatus(serviceInstanceName string) string {
 
 // GetVolumeMountName returns the name of the volume
 func (oc *OcRunner) GetVolumeMountName(dcName string) string {
-	volumeName := CmdShouldPass(oc.path, "get", dcName, "-o", "go-template='"+
+	volumeName := CmdShouldPass(oc.path, "get", "dc", dcName, "-o", "go-template='"+
 		"{{range .spec.template.spec.containers}}"+
 		"{{range .volumeMounts}}{{.name}}{{end}}{{end}}'")
 
@@ -283,7 +283,7 @@ func (oc *OcRunner) GetVolumeMountName(dcName string) string {
 
 // GetVolumeMountPath returns the path of the volume mount
 func (oc *OcRunner) GetVolumeMountPath(dcName string) string {
-	volumePaths := CmdShouldPass(oc.path, "get", dcName, "-o", "go-template='"+
+	volumePaths := CmdShouldPass(oc.path, "get", "dc", dcName, "-o", "go-template='"+
 		"{{range .spec.template.spec.containers}}"+
 		"{{range .volumeMounts}}{{.mountPath}} {{end}}{{end}}'")
